[WIP] First playable tutorial (#1805)
* Hook up input for tutorial

* Set up first run-through

* Tweaks

* Fix lint issues
